#13 Rename module (abc → comparable)
