+++ ndtools/comparable.py
